Implement a simple state loader
To avoid extra dependencies, implement a very
limited and rudimentary state loader of an action
that can save the definition of its user-defined
class for derived class instances.

To facilitate module locations and pickling, an
intermediary JSON file is used to store the plain
text source to be loaded *first* before unpickling
and the location to the pickle file. This will not
be sufficient for complex code layouts, however
this satisfies the need to load an action in a
separate process without previous knowledge about
the file location where the class is defined.

Should the approach need to change, the entire
process' entry points are limited to anything
derived from SaveState and the helper functions.
Thus, changing the internals should remain opaque
to user interactions.