Add ability to run env scripts
This takes the same concept from the lmod python command to output
usable python code to stdout and then exec that code within the same
interpreter/process. To do so, a script is fed into the wrapper script
env_from_script.sh which captures the current environment, sources the
provided script (does NOT run it in a subshell) and compares the
resulting environment. Any differences are applied as `os.environ`
edits, where new or modified values are assigned and missing values
are removed from the python dict.

Both the lmod command and this command are then wrapped in stdout and
stderr redirection to capture any python print statements happening
outside our logging. This then allows us to continue the theme of log
output from stdout coming at the correct time with our logging prefix.