Allow hosts to provide a base env
This base env will be named host name + _env and will always be setup
before any specified environment if this base exists.

Likewise, the base can pass on specific attributes to host envs. By
default, the lmod path and lmod python module are passed down so that
it need only be specified, found, and loaded once.