Fix final reporting of hpc jobs
Actions are now provided to host on pre and post run of all actions.
This allows for careful and deliberate modification of the actions as
seen fit by the host. In the case of HPC hosts, this is used to check
for job completion and final status, thereby allowing waiting for
submitted actions to complete and change there final status from
'submitted' to either 'success' or 'failure'.

The job check is now split into two main checks, complete and status.
Before, it was assumed complete was enough and the status remained
ambiguous as 'submitted'. Now, upon job completion, the status is
queried using another shell command whose output is fed to an abstract
function `check_job_status`. This should then report the final status
of the action as a boolean.