Rework save_state saving mechanism
Previously, save_state would inspect the class (self), and store the
class source code definition of any class outside of the specified
base class using MRO to determine this. This, in theory, works well
to allow the pickled object to be transported virtually anywhere since
the required classes' source code will be packed in the save file.

However, in practice this limits development threefold:
1. All used functions or imports must occur within a class. Code outside
classes will not be captured
2. Imports must utilize absolute names and cannot be imported with
aliases. Likewise, imports of just objects from a module do not work.
3. Imports across python workflow files is not possible as each file
is imported separately into a module and locally stored.

A workaround for 1 does not exist.
A workaround for 2 is to inspect the originating module and capture both
imports and aliases. This does work, but no viable solution was found
for just importing objects (e.g. from matplotlib import pyplot )
A workaround for 3 is to ensure dynamically imported modules are found
in the sys.modules with full module names. This, however, may lead to
confusion on the exact location of where and how to import other modules
within a workflow.

Given all the necessary workarounds explored, the simpler solution is to
dynamically import modules by modifying sys.path. This has been deemed a
sufficiently acceptable replacement for the above constraints.

The added paths will correspond to the user-provided paths to the root
workflow folders, or default ./ otherwise. Found python modules will be
imported as fully named modules from that path (e.g. demo.foobar.workflow
for path workflow/demo/foobar/workflow.py). Paths will be stored as
absolute paths in the sane.user_space module where they will be handed
off to the SaveState class as the default paths to add to sys.path during
loading. Upon loading a save file, sys.path will be modified with these
additional paths, the originating module for the pickled class will be
re-imported (to bring back in any aliases, functions, etc), and then the
object itself unpickled.

This new save paradigm allows for a code development strategy that more
closely feels natural to writing python. It allows users to write
workflows with very little restraints on their python code, vs
previously needing to contain everything to classes. Additionally, it
allows natural compartmentalization of code in and across workflow
folders, with each accessible via the root of the folder as the import
path.

The only constraints now are:
1. Action objects must still be picklable
2. Any used modules, including from the workflow, must be importable on
the action launch side. For hosts executing actions on the same file
system, this is virtually unaffected. However, if this were to be run on
a host that forwards non-local actions across the network or into a
container, the import paths would need to be modified to point to valid
locations for these modules. Thus for user-defined classes, a copy of
some or all of the user workflow folders should be provided in the host
execution space.