Add new virtual_relaunch and virtual_host options
These two options in combination allow for powerful usage of any host
which derives from NonLocalProvider.

On its own, virtual_host just recreates the expected host with a local
set of resources instead of its previous ones, i.e. command-line based
modification of the available resources a host can provide. This is not
very useful when applied by the user, unless the user is within a
"non-local" space (e.g. a compute node and trying to run a workflow)

The virtual_relaunch options, on the other hand, reconstructs the
command line options to re-launch the sane_runner.py as the command of
an action itself. This effectively runs the sane_runner.py again, but
this time executing the workflow as part of the in situ action run. The
resources provided to this virtual_relaunch option are passed over to
the virtual_host option instead on re-launch, and the virtual_relaunch
is cleared to prevent infinitely looping.

Thus, the only action executed is the entire workflow with resource
requirements provided by the virtual_relaunch option. This single action
is launched with the _original_ host, using whatever local or nonlocal
wrapper it wants. Then, the action's run executes the workflow now using
the action's resources as its total local resource available. From there
the workflow is in forced local mode using this virtual_host resource
pool. When applied to an HPC host, for instance, this causes the entire
workflow to execute as a single job with the pool of resources specified
at the command line.