Track origin of JSONConfig objects
Capture the source file of the JSONConfig object instance actual class,
such as host.py or action.py. Then capture the location in the code that
has created this instance. This is generally not useful for config-only
actions that are created from within the orchestrator, even when using
user workflow defined custom classes. However, for JSONConfig objects
created within user workflow registered python functions, this will find
exactly where the instance is created.

Finally, for config loading and patching keep track of every time the
`load_config` call is made passing along an origin. Each time the base
`load_config` is reached, store the origin. For underlying objects
created within config calls, origin is passed to `load_core_config` and
`load_extra_config` as well.

This will then result in a chrononlogically accurate history of the
creation of this JSONConfig object. One would be able to find the class
definition file, then the location of instantiation for any in-code
modifications. Then, in order, see the application of config updates to
the object.

Add the origins field to the actions results for saving