Run action_launcher.py with python3 when not executable
During the creation of .whl packages, the executable attribute of
files is removed for .py files. This is not able to be changed for
files meant to stay in the package directory. All alternatives are for
non-py files, or installs them into the respective environment bin/
folder, which is not conducive to the usage of the action_launcher.

As a workaround, when not running from source (installed or cloned)
the launch command will check if sane/action_launcher.py is executable
and if not will shift the launch command from the action_launcher to
python3 with the action_launcher.py file as the first argument.