Fix the loading of config values
For NonLocalProviders, the underlying mapper of the parent class must be
loaded before any additional internal resources can be added.

For patches:

Note the origin of the patch and number of general types affected. Only
note exact details for more verbose logging.

when processing filters the full config dict must be passed
to each load. This should probably be changed to not modify the config
dict in the load_config() call and instead make a copy inside that
function for internal tracking.

Do not add empty patches for all read JSON files