Add host watchdog
Normally, the orchestrator action processing queue is continued via a
wake event whenever an action completes. This wake event is done by the
actions themselves.

With the introduction of a pending state, we can end up in a situation
where actions are wainting in the queue and the orchestrator is sleeping
but no actions are running in the context of the orchestrator (i.e. a
non-local host is managing them elsewhere like HPC submissions) thus the
orchestrator will never be awoken from an action completing. Since
non-local actions have already completed within the context of the
orchestrator queue they also cannot execute the wake event. A watchdog
function is introduced to allow all hosts to invoke this wake event. As
such, NonLocalProvider hosts can now monitor the progress of actions and
wake the orchestrator when appropriate, changing the respective action
state and status as necessary. This would then allow pending dependency
actions to be re-evaluated, reflecting any progress on its dependencies.