Remove old save file on new run
Due to changes in how cached workflows are saved, the new load was not
working as intended.

Originally, this logic worked when the initial caching feature was made
but now since saves always load in the previous save and recursively
update any changes, simply not loading once is not effective.

To fully force a new run the save file should be deleted. The load()
function in the Orchestrator should most likely be deprecated or at the
very least re-evaluated.